backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase==2.6.0
google-genai
python-multipart
pillow
